ai-services/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.26.2
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
transformers==4.31.0
spacy==3.6.0
nltk==3.8.1
python-jose[cryptography]==3.3.0
requests==2.31.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
gunicorn==21.2.0
python-multipart==0.0.6
pydantic[email]==2.4.2
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pymongo==4.4.1
redis==4.6.0
celery==5.3.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
sentry-sdk==1.32.0 